--- services/catalog/src/simcore_service_catalog/schemas/schemas_dusers.py
@@ -1,41 +1,0 @@
-# TODO: split in InData, OutData
-# TODO: find commons
-# TODO: conversion from/to database's ORM in models.py
-
-from typing import List
-
-from pydantic import BaseModel  # pylint: disable=no-name-in-module
-
-
-class ItemBase(BaseModel):
-    title: str
-    description: str = None
-
-
-class ItemCreate(ItemBase):
-    pass
-
-
-class Item(ItemBase):
-    id: int
-    owner_id: int
-
-    class Config:
-        orm_mode = True
-
-
-class UserBase(BaseModel):
-    email: str
-
-
-class UserCreate(UserBase):
-    password: str
-
-
-class User(UserBase):
-    id: int
-    is_active: bool
-    ditems: List[Item] = []
-
-    class Config:
-        orm_mode = True
